move compareToResult class to own file
