+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/models/fms/MonitoringOrderCompareTo.kt
@@ -2,7 +2,6 @@ package uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.m
 
 import MonitoringOrderChange
 import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.enums.VariationType
-import uk.gov.justice.digital.hmpps.hmppselectronicmonitoringcreateanorderapi.models.fms.config.OrderChangeDetail
 
 fun MonitoringOrder.compareTo(previous: MonitoringOrder): MonitoringOrderCompareToResult {
   val result = MonitoringOrderCompareToResult()
@@ -229,27 +228,3 @@ fun MonitoringOrder.compareTo(previous: MonitoringOrder): MonitoringOrderCompare
 }
 
 class MonitoringOrderCompareToResult : CompareToResult<MonitoringOrderChange>()
-
-open class CompareToResult<T : OrderChangeDetail> {
-  private val _messages = mutableListOf<String>()
-  val messages: MutableList<String> get() = _messages
-
-  private val orderVariationTypes = mutableListOf<VariationType>()
-  val orderVariationType: VariationType
-    get() {
-      return orderVariationTypes.minByOrNull { it.priority } ?: VariationType.OTHER
-    }
-
-  fun addChange(change: T) {
-    _messages += change.message
-    orderVariationTypes += change.orderVariationType
-  }
-
-  fun addOrderVariationType(variationType: VariationType) {
-    orderVariationTypes += variationType
-  }
-
-  fun addMessage(message: String) {
-    _messages += message
-  }
-}
